+++ tests/test_project.py
@@ -35,7 +35,7 @@ def test_project():
     assert len(project.get_mnemonics(['DT'])) == 4
 
     html = project.curve_table_html()
-    assert '<table><tr><th>Idx</th><th>UWI</th><th>Data</th><th>Quality</th>' in html
+    assert '<table><tr><th>Idx</th><th>UWI</th><th>Data</th><th>Passing</th>' in html
     assert "<th>DPHI_SAN</th>" in html
     s = """<td style="background-color:#CCEECC; line-height:80%; padding:5px 4px 2px 4px;">DTS"""
     assert s in html